requirements.txt
# AI 智能教育平台 - 核心依赖

# Streamlit 核心
streamlit>=1.28.0

# 图像处理
Pillow>=10.0.0
PyMuPDF>=1.25.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 环境变量
python-dotenv>=1.0.0

# HTTP 请求
requests>=2.31.0

# AI/ML 相关
openai>=1.45.0

# LangGraph 依赖 (可选)
langgraph>=0.0.20
langchain>=0.1.0
langchain-core>=0.1.0

# Google Gemini SDK (可选)
google-genai>=0.2.0

# 文档处理
python-docx>=1.1.0
openpyxl>=3.1.2
markdown>=3.5.1

# Web 框架 (可选)
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6

# 数据库 (可选)
sqlalchemy>=2.0.0
alembic>=1.13.0